+++ src/source/MVulkanRHI/MVulkanPipeline.cpp
@@ -16,7 +16,7 @@ void MVulkanPipeline::Clean(VkDevice device)
 }
 
 void MVulkanGraphicsPipeline::Create(MVulkanDevice device, VkShaderModule vertShaderModule, VkShaderModule fragShaderModule, VkRenderPass renderPass,
-    PipelineVertexInputStateInfo vertexStateInfo, VkDescriptorSetLayout layout, uint8_t numAttachments)
+    PipelineVertexInputStateInfo vertexStateInfo, VkDescriptorSetLayout layout, uint32_t numAttachments)
 {
     VkPipelineShaderStageCreateInfo vertShaderStageInfo{};
     vertShaderStageInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
